refactor: interfaces, dao package 분리
