+++ src/main/java/racingcar/car/interfaces/Name.java
@@ -1,6 +1,7 @@
-package racingcar.car.model;
+package racingcar.car.interfaces;
 
 public interface Name {
     
     String getValue();
 }
+
