+++ src/main/java/racingcar/car/model/CarPosition.java
@@ -1,6 +1,7 @@
 package racingcar.car.model;
 
 import java.util.Objects;
+import racingcar.car.interfaces.Position;
 
 public final class CarPosition implements Position {
